* Add conditional removal of builds directory

+++ mirror.sh
@@ -3,17 +3,20 @@
 set -e
 
 INDEX="index.json"
+FIRST_URL=true
 
 rm -f "$INDEX"
 curl -s "https://ziglang.org/download/index.json" -o "$INDEX"
-
-rm -rf builds
 
 for URL in $(cat index.json | jq -r '.. | .tarball? | select(. != null)'); do
     echo "url: $URL"
     FILE=$(echo "$URL" | cut -d'/' -f4-)
     echo "file: $FILE"
     if [ ! -f "$FILE" ]; then
+      if [ "$FIRST_URL" = true ]; then
+        rm -rf builds
+        FIRST_URL=false
+      fi
       DIR=$(echo "$URL" | cut -d'/' -f4- | rev | cut -d'/' -f2- | rev)
       echo "dir: $DIR"
       mkdir -p "$DIR"
